requirements.txt
# ============================================================
# Campus Sphere — Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# --- Database ---
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# --- AI / LLM ---
mistralai==1.5.0

# --- RAG System ---
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1

# --- Machine Learning ---
pandas==2.2.3
scikit-learn==1.6.1
numpy==2.2.1

# --- Data Loading ---
openpyxl==3.1.5

# --- Configuration ---
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# --- HTTP Client ---
httpx==0.28.1

# --- CORS ---
# (included in FastAPI)

# --- Testing ---
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# --- Utilities ---
loguru==0.7.3
